HCS-4: add investment entry policy and add authorize in crud

+++ app/Http/Livewire/Investments/Entries/Destroy.php
@@ -4,16 +4,21 @@ namespace App\Http\Livewire\Investments\Entries;
 
 use App\Models\{Investment, InvestmentEntry};
 use Illuminate\Contracts\View\View;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Livewire\Component;
 
 class Destroy extends Component
 {
+    use AuthorizesRequests;
+
     public ?Investment $investment = null;
 
     public ?InvestmentEntry $investmentEntry = null;
 
     public function save(): void
     {
+        $this->authorize('delete', $this->investmentEntry);
+
         $this->investmentEntry->delete();
 
         $this->emit('investment::entry::deleted');
